Added support for other elements of the watershed graph

+++ app/src/main/java/com/viableindustries/waterreporter/TerritoryActivity.java
@@ -571,7 +571,7 @@ public class TerritoryActivity extends AppCompatActivity implements TimelineFilt
 
         try {
 
-            int mTerritoryId = mTerritory.id;
+            long mTerritoryId = mTerritory.id;
 
         } catch (NullPointerException e) {
